restructure project to align more with convention
