+++ .gitignore
@@ -16,3 +16,4 @@
 .vscode
 
 node_modules/
+debug
