--- model.go
@@ -1,95 +1,0 @@
-package main
-
-// Success response
-// swagger:response ok
-type swaggScsResp struct {
-	// in:body
-	Body struct {
-		// HTTP Status Code 200
-		Code int `json:"code"`
-	}
-}
-
-// Boolean response
-// swagger:response bool
-type swaggBoolResp struct {
-	// in:body
-	Body struct {
-		// HTTP Status Code 200
-		Code int `json:"code"`
-		// Boolean true/false
-		Data bool `json:"data"`
-	}
-}
-
-// Error Bad Request
-// swagger:response badReq
-type swaggErrBadReq struct {
-	// in:body
-	Body struct {
-		// HTTP status code 400 - Status Bad Request
-		Code int `json:"code"`
-		// Detailed error message
-		Message string `json:"message"`
-	}
-}
-
-// Error Forbidden
-// swagger:response forbidden
-type swaggErrForbidden struct {
-	// in:body
-	Body struct {
-		// HTTP status code 403 - Forbidden
-		Code int `json:"code"`
-		// Detailed error message
-		Message string `json:"message"`
-	}
-}
-
-// Error Not Found
-// swagger:response notFound
-type swaggErrNotFound struct {
-	// in:body
-	Body struct {
-		// HTTP status code 404 - Not Found
-		Code int `json:"code"`
-		// Detailed error message
-		Message string `json:"message"`
-	}
-}
-
-// Error Conflict
-// swagger:response conflict
-type swaggErrConflict struct {
-	// in:body
-	Body struct {
-		// HTTP status code 409 - Conflict
-		Code int `json:"code"`
-		// Detailed error message
-		Message string `json:"message"`
-	}
-}
-
-// Error Interval Server
-// swagger:response internal
-type swaggErrInternal struct {
-	// in:body
-	Body struct {
-		// HTTP status code 500 - Internal server error
-		Code int `json:"code"`
-		// Detailed error message
-		Message string `json:"message"`
-	}
-}
-
-// A ValidationError is a swagger response to represent error
-//
-// swagger:response validationError
-type ValidationError struct {
-	// in: body
-	Body struct {
-		Code    int32  `json:"code"`
-		Message string `json:"message"`
-		Field   string `json:"field"`
-	} `json:"body"`
-}
